khinsider tool - Windows version!
Cloudflare protection really does not like Windows users, flagging any requests from them as bot activity and returning 403 errors, regardless of what is used for the request header. To circumvent this, we need to install the cloudscraper package, which is mostly a drop-in replacement, though it doesn't have ConnectionError like requests have, so clauses checking the connection may not work/throw up appropriate errors going forward.

Depending on whether Linux/Mac users similarly are treated as bot activity, this Windows version may merge with the OG khinsider in future.